added necessary items to allow for use as a package
